Simplify `syntax-highlighting` example
