--- examples/syntax-highlighting/gatsby-config.js
@@ -1,19 +1,0 @@
-module.exports = {
-  plugins: [
-    {
-      resolve: 'gatsby-source-filesystem',
-      options: {
-        name: 'pages',
-        path: __dirname + '/src/pages/'
-      }
-    },
-    {
-      resolve: 'gatsby-plugin-mdx',
-      options: {
-        defaultLayouts: {
-          default: require.resolve('./src/components/Layout')
-        }
-      }
-    }
-  ]
-}
