+++ examples/syntax-highlighting/next.config.js
@@ -1,0 +1,5 @@
+const withMDX = require('@next/mdx')()
+
+module.exports = withMDX({
+  pageExtensions: ['js', 'jsx', 'mdx']
+})
