+++ examples/syntax-highlighting/package.json
@@ -3,17 +3,18 @@
   "version": "2.0.0-next.9",
   "private": true,
   "scripts": {
-    "build": "gatsby build",
-    "start": "gatsby develop"
+    "build": "next build",
+    "dev": "next",
+    "start": "next start"
   },
   "dependencies": {
+    "@mdx-js/loader": "2.0.0-next.9",
     "@mdx-js/mdx": "2.0.0-next.9",
-    "@mdx-js/react": "2.0.0-next.9",
-    "gatsby": "^2.0.0",
-    "gatsby-plugin-mdx": "^1.0.0",
-    "gatsby-source-filesystem": "^2.0.0",
+    "@next/mdx": "^10.0.0",
+    "next": "^11.0.0",
     "prism-react-renderer": "^1.0.0",
-    "react-live": "^2.0.0",
-    "unified-ui": "^0.0.3"
+    "react": "^17.0.0",
+    "react-dom": "^17.0.0",
+    "react-live": "^2.0.0"
   }
 }
