+++ examples/syntax-highlighting/pages/index.mdx
