+++ examples/syntax-highlighting/readme.md
@@ -1,4 +1,4 @@
-# Gatsby + MDX
+# Syntax highlighting
 
 ```sh
 yarn && yarn start
